feat: ダンジョン難易度1〜10の再設計とドロップ配線、スレッド/装備曲線の是正
難易度を倍率キーで表現できるようにし、討伐素材とスレッドのフィールドドロップを
配線し、スレッドが低帯だけ極端に強くなる曲線を割合系へ振り替えた。

- 倍率キー2種を新設（max-health-multiplier / attack-power-multiplier）。
  層をまたいで掛け合わさるため default × dungeon × mob が積になる。
  per-mob の絶対値 attack-power があると mergeAttack がスコープ倍率を捨てる。
- yml だけでは書けなかった3要件を Java のキーとして追加
  chance-by-level: lv1→lv100 の連続補間。帯境界の段差が消えて 102→23 エントリに減った
  where: field|dungeon|any。mob-level-table.yml の dungeon-only はファイル単位の
    1本トグルで field-only が書けなかった
  baby: Ageable#isAdult。MobTargetFilter は EntityType と mobIds の2軸しか無かった
- 実数ダメージ系ステ（attack-power / flat-bonus-damage / bleed-damage）を
  スレッド枠から全廃し割合系へ振替。TTK ぶれ 65.2倍 → 6.1倍（理論下限 6.5倍）。
  案A（use-level-requirement で帯分割）は成立しない。use-level の強制は
  UseRequirementListener と ArmorUseGateListener の2経路だけで、
  ソケット装着はどちらも発火しないため書いても無言の no-op になる。
- socketed-only-stats を新設し「装備に挿さず手に持つだけで全ステが乗る」穴を塞いだ。
  PlayerStatAggregator.isWornOnlyArmor はホワイトリストではなくブラックリストで、
  どの防具スロットにも解決しないスレッドが default -> false に落ちていた。
  DerivedItemStats.resolve/resolveMultipliers の合流点8箇所すべてに門を通している。
- 帯目標ガードを threads.yml の stackable/max と thread-sets.yml の累積規則から
  導出する形へ組み直した。stackable/max は「1つの防具あたり」の上限であって
  全体上限ではなく、mana_regen は max:3 ＝防具4部位だけで12本挿さる。
  割合ダメージを配るセットは hero_of_the_village の1本へ集約した
  （mana_regen→magic-resistance、night_vision→dodge-chance。会心系にしないのは
  意図的で、crit は実DPSを押し上げるのに帯ガードが数えないため）。
- item-stats.yml のスレッド6種（匠/儀式師/潮読み/選書/調香/鑑識）で advanced: が
  丸ごと落ち、主ステの random まで min:0 に化けていたのを復元。

実測: TrinityForge 3946 / 26 failed / 2 skipped、config-editor 1237 / 22 failed / 1 skipped。
いずれも新規失敗ゼロ。クリーン worktree では自分のパスのみで 9 failed
（HEAD も 9 だが中身が違い、warden_tendril 系5件と綴り1件を解消している）。

注意: 本コミットには別セッションの未コミット変更も含まれる（ユーザー判断）。
collection.yml は config-editor の保存事故を含んでおり、structure カテゴリが
丸ごと消え、infinity から abyss_* / binder_* が消えている。

+++ TrinityForge/src/main/java/com/trinityforge/mob/FocusHpDisplay.java
@@ -293,9 +293,13 @@ public final class FocusHpDisplay implements Listener {
         MobData data = MobData.of(target);
         int level = data.hasProfile() ? data.level() : 0;
         Component nameComponent = displayNames.displayName(target);
-        int curHp = (int) Math.ceil(target.getHealth());
+        // 2026-08-14: int だと double→int の narrowing が Integer.MAX_VALUE で「飽和」する(JLS 5.1.3。
+        // 負値に巻き戻りはしないが 2,147,483,647 に張り付いて "2.15B" と嘘の値を出す)。現状の最大HP
+        // 69,371,755 は int の範囲内なので今すぐ壊れてはいないが、HP倍率は層をまたいで掛け合わさる
+        // (default × dungeon × mob)ので上限が読めない。long にしておけば 9.2e18 まで飽和しない。
+        long curHp = (long) Math.ceil(target.getHealth());
         AttributeInstance maxHealthAttr = target.getAttribute(Attribute.MAX_HEALTH);
-        int maxHp = maxHealthAttr != null ? (int) Math.ceil(maxHealthAttr.getValue()) : curHp;
+        long maxHp = maxHealthAttr != null ? (long) Math.ceil(maxHealthAttr.getValue()) : curHp;
         // どちらの型に寄って耐性/防御が設定されているかを示す任意タグ(依頼2)。既存の
         // combat/mob-defaults.yml 系キー(PDCへ既に焼かれている)から導出するだけで、新しい設定面は
         // 増やさない — MobData#defenseFor はプロファイル無しモブでは全項目0を返すので、その場合は
